Support forge 1.16.5. Closes #58

+++ tests/src/loaders/forge.ts
@@ -1,7 +1,8 @@
+import {exists} from '@std/fs'
 import { parse as parseXml } from '@libs/xml'
 import { MAX_HEAP_SIZE, ServerType } from '../const.ts'
 import { Server } from '../server.ts'
-import { downloadFile } from '../utils.ts'
+import {downloadFile, JAVA_8_PATH} from '../utils.ts'
 import { basename } from '@std/path'
 import { ServerInstallError } from '../errors.ts'
 
@@ -38,10 +39,29 @@ abstract class AnyForgeServer extends Server {
         `Failed to install ${this.type} ${this.version}`,
       );
     }
+
+    const javaPath = this.version == '1.16.5' ? JAVA_8_PATH : 'java'
+
     const runShFile = new URL('run.sh', this.serverDir);
-    let runShContent = await Deno.readTextFile(runShFile);
-    runShContent = runShContent.replace('java', 'exec java');
-    await Deno.writeTextFile(runShFile, runShContent);
+    if (!await exists(runShFile)) {
+      let forgeJar = null;
+      for await (const entry of Deno.readDir(this.serverDir)) {
+        if (
+          entry.isFile && entry.name.startsWith('forge-') &&
+          entry.name.endsWith('.jar') && !entry.name.includes('installer')
+        ) {
+          forgeJar = entry.name;
+        }
+      }
+      let runShContent = '#!/usr/bin/env sh\n';
+      runShContent += `exec ${javaPath} -jar ${forgeJar} "$@"\n`;
+      await Deno.writeTextFile(runShFile, runShContent);
+      await Deno.chmod(runShFile, 0o755);
+    } else {
+      let runShContent = await Deno.readTextFile(runShFile);
+      runShContent = runShContent.replace('java', `exec ${javaPath}`);
+      await Deno.writeTextFile(runShFile, runShContent);
+    }
   }
 
   protected override get startCommand() {
@@ -60,18 +80,21 @@ abstract class AnyForgeServer extends Server {
   }
 
   protected override getDependencies() {
-    if (this.version.split('.')[1] == '21' || this.version == '1.20.6') {
+    const versionMinor = parseInt(this.version.split('.')[1])
+    if (versionMinor == 21 || this.version == '1.20.6') {
       return [
         'https://cdn.modrinth.com/data/ordsPcFz/versions/uZAwdzAR/kotlinforforge-5.7.0-all.jar',
       ];
-    } else if (this.version.split('.')[1] == '20' || this.version == '1.19.4') {
+    } else if (versionMinor == 20 || this.version == '1.19.4') {
       return [
         'https://cdn.modrinth.com/data/ordsPcFz/versions/hmeyC41q/kotlinforforge-4.11.0-all.jar',
       ];
-    } else if (['19', '18'].includes(this.version.split('.')[1])) {
+    } else if ([19, 18].includes(versionMinor)) {
       return [
         'https://cdn.modrinth.com/data/ordsPcFz/versions/NBn3sEQk/kotlinforforge-3.12.0-all.jar',
       ];
+    }  else if (versionMinor == 16) {
+      return [];
     } else {
       throw new Error(`Unsupported forge version ${this.version}`);
     }
